refactor: move CLI entrypoint from cli-main.ts to index.ts and update bin path
