+++ src/cli/index.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { runCli } from "./cli";
+
+runCli(process.argv);
